requirements.txt
pyrofork==2.2.11
tgCrypto
python-dotenv
myjdapi